Update README.md
Use reference-style links.

Don't use a promptinit-ish code anymore, as we don't want to encourage
the use of it, mainly because: "prompt_opts is evil", prompt_cleanup
does not do a complete cleanup, and it's overkill for the daily
I-just-use-a-single-prompt-during-my-whole-shell-session-that-took-me-a-good-while-to-configure-to-my-liking
scenario.

Mention more explicitly that verbose mode can be slower, without getting
into too much detail, as the speed of the respective git commands vary
a lot depending on the repository and the git version being used.
See zimfw/zimfw#344