Update git-action based on libgit2
romkatv/gitstatus is also using git actions based on libgit2.
And we want to have a separate gitstatus-info module that is like this
git-info but depending on gitstatus instead of git.
Our implementation and list of detected actions was even outdated when
compared to the current vcs_info one. So trying another direction going
forward...