🛷  Refactor git detection (rev-parse to zsh impl)
The current implementation always calls git rev-parse.
This means `git` will always be invoked, even if there is
no git directory.

To demonstrate, zmodload zsh/zprof top of zshrc and run
zprof (manually) after shell invocation.

If you have a phat home directory, the invocation of git-info can be
seen consuming a fair bit of clock time. I don't think this actually
fixes the "problem" of git-info consuming clock time, but it does reduce
the number of instances `git rev-parse` will be invoked.

This implementation may also be buggy, or there may be a reason we're
relying on the git rev-parse implementation that I don't recall.

Signed-off-by: Matt Hamilton <m@tthamilton.com>