Be faster by running git commands in parallel \o/
This makes git-info run almost 3 times faster. Uses MULTIOS with
multiple command substitutions to run all commands asynchronously.

This does not apply to the git status command in verbose mode, because
the command outputs multiple lines.

Also removed the git rev-parse command, to check if we're inside a git
repo, in favor of just running the git symbolic-ref command, as we need
to get the branch name synchronously anyway.